chore : format align

+++ src/KabiswapERC20.sol
@@ -1,6 +1,6 @@
 pragma solidity >=0.8.10;
 
-import '../interfaces/IERC20.sol';
+import "../interfaces/IERC20.sol";
 
 contract KabiswapERC20 is IERC20 {
     string public constant _name = "Kabiswap";
@@ -8,22 +8,22 @@ contract KabiswapERC20 is IERC20 {
     uint8 public constant decimals = 18;
     uint256 public _totalSupply;
 
-    mapping(address => uint256) nonces; 
-    mapping(address => uint) balances;
-    mapping(address => mapping(address => uint)) _allowance;
+    mapping(address => uint256) nonces;
+    mapping(address => uint256) balances;
+    mapping(address => mapping(address => uint256)) _allowance;
 
     bytes32 public DOMAIN_SEPARATOR = keccak256(
-            abi.encode(
-                keccak256('EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)'),
-                keccak256(bytes(_name)),
-                keccak256(bytes('1')),
-                block.chainid,
-                address(this)
-            )
-        );
-    
+        abi.encode(
+            keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"),
+            keccak256(bytes(_name)),
+            keccak256(bytes("1")),
+            block.chainid,
+            address(this)
+        )
+    );
+
     constructor() {
-        _totalSupply = 1_000_000*(10**decimals);
+        _totalSupply = 1_000_000 * (10 ** decimals);
     }
 
     function mint() external payable {
@@ -36,7 +36,7 @@ contract KabiswapERC20 is IERC20 {
         balances[msg.sender] -= amount;
     }
 
-    function totalSupply() external view returns (uint256){
+    function totalSupply() external view returns (uint256) {
         return _totalSupply;
     }
 
@@ -44,13 +44,13 @@ contract KabiswapERC20 is IERC20 {
         return balances[account];
     }
 
-    function _transfer(address from, address to, uint value) internal {
+    function _transfer(address from, address to, uint256 value) internal {
         require(balances[from] >= value, "Insufficient value.");
         balances[from] -= value;
         balances[to] += value;
     }
 
-    function transfer(address to, uint value) external returns (bool) {
+    function transfer(address to, uint256 value) external returns (bool) {
         _transfer(msg.sender, to, value);
         return true;
     }
@@ -59,17 +59,17 @@ contract KabiswapERC20 is IERC20 {
         return _allowance[owner][spender];
     }
 
-    function _approve(address owner, address spender, uint value) internal {
+    function _approve(address owner, address spender, uint256 value) internal {
         require(owner != address(0) || spender != address(0), "Zero address is not allowed.");
         _allowance[owner][spender] = value;
     }
 
-    function approve(address spender, uint value) external returns (bool) {
+    function approve(address spender, uint256 value) external returns (bool) {
         _approve(msg.sender, spender, value);
         return true;
     }
 
-    function transferFrom(address from, address to, uint value) external returns (bool) {
+    function transferFrom(address from, address to, uint256 value) external returns (bool) {
         require(_allowance[from][msg.sender] > 0, "Not approved.");
         _transfer(from, to, value);
         return true;
@@ -83,30 +83,28 @@ contract KabiswapERC20 is IERC20 {
         return _symbol;
     }
 
-    function permit(
-        address owner, 
-        address spender, 
-        uint value, 
-        uint deadline, 
-        uint8 v, 
-        bytes32 r, 
-        bytes32 s
-        ) external {
-            require(deadline >= block.timestamp, 'deadline is EXPIRED.');
-            bytes32 digest = keccak256(
-                abi.encodePacked(
-                    '\x19\x01', 
-                    DOMAIN_SEPARATOR, 
-                    keccak256(abi.encode(
+    function permit(address owner, address spender, uint256 value, uint256 deadline, uint8 v, bytes32 r, bytes32 s)
+        external
+    {
+        require(deadline >= block.timestamp, "deadline is EXPIRED.");
+        bytes32 digest = keccak256(
+            abi.encodePacked(
+                "\x19\x01",
+                DOMAIN_SEPARATOR,
+                keccak256(
+                    abi.encode(
                         keccak256("Permit(address owner,address spender,uint256 value,uint256 nonce,uint256 deadline)"),
                         owner,
                         spender,
-                        value, nonces[owner]++,
+                        value,
+                        nonces[owner]++,
                         deadline
-                        ))
-                    ));
-            address recoveredAddress = ecrecover(digest, v, r, s);
-            require(recoveredAddress != address(0) && recoveredAddress == owner, "Invalid Signature.");
-            _approve(owner, spender, value);
-        }
+                    )
+                )
+            )
+        );
+        address recoveredAddress = ecrecover(digest, v, r, s);
+        require(recoveredAddress != address(0) && recoveredAddress == owner, "Invalid Signature.");
+        _approve(owner, spender, value);
+    }
 }
